Provides a list of numbers that divide a given number
Provides all of the positive integers that divide some given integer
into an integer. The code is written in c, and it uses a is_prime() and
pool_of_primes() algorithm such that we return a pool of primes that
divide our number, and we use this pool accompanied with bitwise logic
and a combinations algorithm to provide all of the integers (rather than
a for loop).